move convert function to client
