+++ Composer/packages/client/src/utils/convertUtils/parsePathToFocused.ts
