+++ Composer/packages/client/src/utils/convertUtils/parsePathToSelected.ts
