+++ Composer/packages/lib/shared/src/index.ts
@@ -12,5 +12,4 @@ export * from './promptTabs';
 export * from './appschema';
 export * from './types';
 export * from './constant';
-export * from './convertUtils';
 export { LgMetaData, LgTemplateRef } from '@bfc/indexers';
